refactor. [Domain] 기존 data와 domain으로 나뉘어 있던 디렉토리 구조 변경
