+++ lib/src/domain/helpers/api_result.dart
@@ -1,0 +1,15 @@
+enum ApiErrors {
+  unexpected,
+  network,
+  noResult,
+  invalidInput,
+}
+
+class ApiResult<T> {
+  final bool success;
+  final ApiErrors? error;
+  final String? msg;
+  final T? value;
+
+  ApiResult(this.success, {this.error, this.msg, this.value});
+}
