--- lib/src/domain/models/models.dart
@@ -1,5 +1,0 @@
-export 'brick.dart';
-export 'hashtag.dart';
-export 'user.dart';
-
-export 'result.dart';
