--- lib/src/domain/repositories/repositories.dart
@@ -1,2 +1,0 @@
-export 'brick_repository.dart';
-export 'user_repository.dart';
