+++ lib/src/domain/user/models/user.dart
